tests clean up
